feat: add processing for the result of creating a new shopping list

+++ app/src/main/java/alex/dev/shoppingnotelistapp/data/data_source/local/room/dao/NoteItemDao.kt
@@ -29,6 +29,6 @@ interface NoteItemDao {
 
     //Получение заметки по Id
     @Query("SELECT * FROM note_items WHERE note_id = :noteId")
-    suspend fun getNoteItemById(noteId: Int): NoteItem
+    suspend fun getNoteItemById(noteId: Long): NoteItem
 
 }
